Updated working functionality for CustomComponent
Updated the CircuitXmlToFormulaConversion Service to handle cases like
disconnected wires and gates.
Improved calculation efficiency by using the ^ sign for XOR in boolean
expressions instead of converting to simpler gates.
Code was tested for a Full Adder, a simple OR gate, a combination of 4
different gates. Got 100% accuracy on all of them.