fix: wire drawing bug fixes (#19)
fix: wires not commit on the first click

After commiting a wire, the second wire would not commit because
the isCornerPointAdded variable was not being reset, fixed it
by adding a single line in HandleWireCommit function

fix: redundant ellipses left after wire is deleted

The redundant ellipses were left because they were being added
to canvas directly. Removed them from the canvas with one line
of code to fix the issue.