fix: issues with components & terminals overlapping
wires were overlapping with components due to wrong
implementation of a function and with terminals due to a missing
check for terminals overlapping after the corner point was being
added.

fix: wire placement control with single-click (#20)

wires can now be committed with the single (left) click when
over a terminal, and with a double (left) click when not on a
terminal.

addit fix: adding the cursor with corner point

Previously, doing a click would only add the corner point, now
it also adds the cursor point to the points.

patch: remove condition 9 from wires

this fixes the false positive signal for now but needs to be
fixed properly later